+++ libraries/stdlib/unsigned/src/kotlin/UInt.kt
@@ -148,10 +148,10 @@ public inline class UInt @PublishedApi internal constructor(@PublishedApi intern
 
     /** TODO */
     @kotlin.internal.InlineOnly
-    public inline fun mod(other: UByte): UInt = this.mod(other.toUInt())
+    public inline fun mod(other: UByte): UByte = this.mod(other.toUInt()).toUByte()
     /** TODO */
     @kotlin.internal.InlineOnly
-    public inline fun mod(other: UShort): UInt = this.mod(other.toUInt())
+    public inline fun mod(other: UShort): UShort = this.mod(other.toUInt()).toUShort()
     /** TODO */
     @kotlin.internal.InlineOnly
     public inline fun mod(other: UInt): UInt = rem(other)
